dedup: extract duplicated helpers + add dupehound CI
Scripts:
- gen-skill-docs.ts: import extractNameAndDescription and
  condenseOpenAIShortDescription from scripts/resolvers/codex-helpers
  instead of carrying local copies.

Browse:
- platform.ts: add validateOutputPath + SAFE_OUTPUT_DIRECTORIES.
- meta-commands.ts, write-commands.ts: import from platform (was
  duplicated, identical 7-line body).
- shorten.ts: new shared path-shortening helper.
- server.ts, sidebar-agent.ts: wrap import (was duplicated, identical
  8-line body using BROWSE_BIN/B).

CI:
- .github/workflows/dupehound.yml: scan (slop score) and check
  (--diff against base) jobs. check is continue-on-error for now;
  promote to gate after FP rate <10% over ~10 PRs.

Slop score: 1.9% -> 1.8% (681 -> 623 deletable lines, 25 -> 21
clusters). Cluster 3 (captureBasicData in extension/) skipped:
content.js and inspector.js run in different isolated worlds and
need chrome.scripting.executeScript changes to share code.
