+++ .github/workflows/dupehound.yml
@@ -1,0 +1,76 @@
+name: dupehound
+
+on:
+  pull_request:
+    paths:
+      - 'browse/**'
+      - 'design/**'
+      - 'scripts/**'
+      - 'extension/**'
+      - 'bin/**'
+      - 'agents/**'
+      - 'autoplan/**'
+  push:
+    branches: [main]
+    paths:
+      - 'browse/**'
+      - 'design/**'
+      - 'scripts/**'
+      - 'extension/**'
+      - 'bin/**'
+      - 'agents/**'
+      - 'autoplan/**'
+
+permissions:
+  contents: read
+
+jobs:
+  check:
+    name: Block new duplicates
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    # Set to false once FP rate is proven <10% over ~10 PRs
+    continue-on-error: true
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Install dupehound
+        run: |
+          curl -sL https://github.com/Rafaelpta/dupehound/releases/latest/download/dupehound-x86_64-unknown-linux-gnu.tar.gz | tar xz
+          sudo mv dupehound /usr/local/bin/
+
+      - name: Block new duplicates vs base
+        env:
+          PR_BASE: ${{ github.event.pull_request.base.ref }}
+        run: |
+          if [ -n "$PR_BASE" ]; then
+            dupehound check --diff "origin/$PR_BASE" .
+          else
+            # push event — diff against HEAD~1
+            dupehound check --diff HEAD~1 .
+          fi
+
+  scan:
+    name: Repo slop score
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Install dupehound
+        run: |
+          curl -sL https://github.com/Rafaelpta/dupehound/releases/latest/download/dupehound-x86_64-unknown-linux-gnu.tar.gz | tar xz
+          sudo mv dupehound /usr/local/bin/
+
+      - name: Scan and post comment with current score
+        run: |
+          dupehound scan . | tee /tmp/dh-scan.txt
+          SCORE=$(grep -oE 'SLOP SCORE[[:space:]]+[0-9.]+%' /tmp/dh-scan.txt | head -1 | grep -oE '[0-9.]+%' || echo "n/a")
+          echo "## dupehound slop score" >> $GITHUB_STEP_SUMMARY
+          echo "**${SCORE}**" >> $GITHUB_STEP_SUMMARY
